Fixed a problem where a !isset was used to detect if the navButtons option was not set.
However this option was always being set by the get_form_html method in the dynamic.php file but could be left blank. So have changed the test to check for empty instead of !isset.

This fixed an issue where the Next Step and Prev Step buttons were broken in Wizard mode and not responding at all.